Moved history into telegram and findTelUsers into utils.
