+++ crawler/utils.py
@@ -3,6 +3,7 @@
 #Aethra/crawler/utils.py
 
 from dbAccess import dbAccess
+import datetime
 import videohash2
 import shutil
 import time
@@ -65,3 +66,107 @@ class utils:
             print(cutPath)
 
         return videoHashHex, videoHashDec
+    
+
+    #---------------------------------------------------------------------------
+    def resolve(self, channel, post):
+
+        """
+        Desc: Resolves if a post should be added to the queue.
+        
+        Input:
+            - post: object from snscrape.
+        """
+
+        #First checking for mentioned accounts that we might not have
+        #stored yet.
+        self.findTelUsers(post.outlinks)
+
+        lastSlash = post.url.rfind('/')
+        postID = post.url[lastSlash+1:]
+        url = f"https://t.me/{channel}/{postID}?single"
+
+        #Checking to see if the post contains a video.
+        try:
+            vidLength = videohash2.video_duration(url=url)
+
+        except videohash2.exceptions.DownloadFailed as err:
+            print(f"[{datetime.datetime.now()}] Caught: {err}",
+                "continuing.")
+            return err
+
+        except videohash2.exceptions.FFmpegFailedToExtractFrames as err:
+            print(f"[{datetime.datetime.now()}] Caught: {err}",
+                "continuing.")
+            return err
+
+        if vidLength > 300:
+            return "tooLong"
+        
+        postInQ = self.dba.findQPost("telegram", channel, postID)
+        print("Post in Queue:", postInQ)
+
+        if not postInQ:
+            postExists = self.dba.postCheck("telegram",
+                                            channel, postID)
+
+            if not postExists:
+
+                postOb = self.postOb()
+                postOb.platform = "telegram"
+                postOb.author = channel
+                postOb.id = postID
+                postOb.text = post.content
+                postOb.timestamp = time.time()
+                postOb.uTime = datetime.datetime.timestamp(post.date)
+                postOb.vidLength = vidLength
+
+                print(post)
+
+                self.dba.addQPost(postOb)
+
+
+    #---------------------------------------------------------------------------
+    def findTelUsers(self, outlinks):
+
+        """
+        Desc: Finds new users in the outlinks of telegram posts.
+        """
+
+        print(f"[{datetime.datetime.now()}] findTelUsers()")
+
+        for link in outlinks:
+
+            if link.find("https://t.me/s/+") != -1:
+                continue
+
+            elif link.find("https://t.me/+") != -1:
+                continue
+        
+            elif link.find("https://t.me/s/") != -1:
+                telURL = "https://t.me/s/"
+                
+            elif link.find("https://t.me/") != -1:
+                telURL = "https://t.me/"
+            
+            else:
+                continue
+
+
+            nameStart = link.rfind(telURL) + len(telURL)
+
+            if link[len(telURL):].find("/") == -1 and link.find("?") == -1:
+                username = link[nameStart:]
+
+            elif link.find("?") == -1:
+                nameEnd = link.rfind("/")
+                username = link[nameStart:nameEnd]
+
+            else:
+                continue
+            
+
+            userExists = self.dba.findTelUser(username)
+            if not userExists:
+                print(f"[{datetime.datetime.now()}] {username} added to telegram list.")
+                self.dba.addTelUser(username)
